fix: - some issue with app.prop - add case without profiles - NPE

+++ src/main/java/com/liquibase/repositories/AuthorDao.java
@@ -4,9 +4,9 @@ import com.liquibase.entities.Author;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 
-public interface AuthorDao extends JpaRepository<Author,Integer> {
+public interface AuthorDao extends JpaRepository<Author, Long> {
 
-//    @Query("SELECT u FROM Author u WHERE u.name = ?1")
+    //    @Query("SELECT u FROM Author u WHERE u.name = ?1")
     @Query("select u from #{#entityName} u where u.name = ?1")
     Author getByName(String authorName);
