+++ src/main/java/com/liquibase/repositories/EmployeeDao.java
@@ -3,6 +3,6 @@ package com.liquibase.repositories;
 import com.liquibase.entities.Employee;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface EmployeeDao extends JpaRepository<Employee,Integer> {
+public interface EmployeeDao extends JpaRepository<Employee, Long> {
 
 }
