+++ src/main/java/com/liquibase/repositories/ProjectDao.java
@@ -3,6 +3,6 @@ package com.liquibase.repositories;
 import com.liquibase.entities.Project;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ProjectDao extends JpaRepository<Project,Integer> {
+public interface ProjectDao extends JpaRepository<Project, Long> {
 
 }
